requirements.txt
Flask
supabase
python-dotenv
pandas
openpyxl
gunicorn
fpdf
